Move toolbar fiters type to it's own files
