+++ framework/index.ts
@@ -43,8 +43,8 @@ export * from './PageNav/PageNavSidebar';
 export * from './PageTable/PageTable';
 export * from './PageTable/PageTableCards';
 export * from './PageTable/PageTableColumn';
-export * from './PageTable/PageToolbar';
-export * from './PageTable/PageToolbarFilter';
+export * from './PageTable/PageToolbar/PageToolbar';
+export * from './PageTable/PageToolbar/PageToolbarFilter';
 export * from './PageTable/useTableItems';
 export * from './PageTabs';
 export * from './Settings';
