+++ framework/useInMemoryView.tsx
@@ -1,7 +1,7 @@
 import getValue from 'get-value';
 import { useEffect, useMemo } from 'react';
 import { ITableColumn } from './PageTable/PageTableColumn';
-import { IToolbarFilter } from './PageTable/PageToolbarFilter';
+import { IToolbarFilter } from './PageTable/PageToolbar/PageToolbarFilter';
 import {
   ISelected,
   useFiltered,
