+++ frontend/eda/Resources/decision-environments/DecisionEnvironments.tsx
@@ -1,7 +1,7 @@
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 import { PageHeader, PageLayout, PageTable } from '../../../../framework';
-import { PageTableViewTypeE } from '../../../../framework/PageTable/PageTableViewType';
+import { PageTableViewTypeE } from '../../../../framework/PageTable/PageToolbar/PageTableViewType';
 import { RouteObj } from '../../../Routes';
 import { API_PREFIX } from '../../constants';
 import { EdaDecisionEnvironment } from '../../interfaces/EdaDecisionEnvironment';
